refacto MOMERepertoire + rewrite MOME + clean notebook
